% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForGithub.R
\name{squrt}
\alias{squrt}
\title{Calculate X to the power of 1/Y.}
\usage{
squrt(x, y)
}
\arguments{
\item{x}{The input, which is a number that
you would like to calculate the square root for.}

\item{y}{The input, which is also a number that the function takes
as the exponent.
as the exponent.}
}
\description{
Calculate X to the power of 1/Y.
}
\examples{
squrt(25,2)
}
